# backend/src/option_arb/api/opportunities.py
from __future__ import annotations

from typing import Any

from fastapi import APIRouter, HTTPException, Query
from sqlmodel import select

from option_arb.api.schemas import (
    ErrorResponse,
    OpportunityResponse,
    OpportunitySortBy,
    OpportunityStatsResponse,
    SortDirection,
)
from option_arb.config import Network
from option_arb.db.models import Opportunity, OpportunityStatus
from option_arb.db.session import get_session

router = APIRouter(prefix="/api/opportunities", tags=["opportunities"])


@router.get("/stats", response_model=list[OpportunityStatsResponse])
async def opportunity_stats(
    days: int = Query(default=30, ge=1, le=365),
    symbol: str | None = None,
    network: Network | None = None,
) -> list[dict[str, Any]]:
    """Opportunités agrégées par paire d'exchanges sur N jours."""
    from datetime import UTC, datetime, timedelta

    import sqlalchemy as sa

    since = datetime.now(UTC) - timedelta(days=days)
    clauses = ["detected_at >= :since", "status != 'REJECTED'"]
    params: dict[str, Any] = {"since": since}
    if symbol:
        clauses.append("symbol = :symbol")
        params["symbol"] = symbol
    if network:
        clauses.append("network = :network")
        params["network"] = network
    where = " AND ".join(clauses)
    sql = sa.text(
        f"""
        SELECT
            buy_from,
            sell_to,
            COUNT(*)                                                                          AS count,
            SUM(net_profit_usd)                                                               AS total_net_profit_usd,
            SUM(fees_usd)                                                                      AS total_fees_usd,
            AVG(apr_pct)                                                                      AS avg_apr_pct,
            MAX(net_profit_usd)                                                               AS best_net_profit_usd
        FROM opportunities
        WHERE {where}
        GROUP BY buy_from, sell_to
        ORDER BY total_net_profit_usd DESC
        """
    )

    async with get_session() as sess:
        rows = list((await sess.execute(sql, params)).mappings())

    return [
        {
            "buy_from": r["buy_from"],
            "sell_to": r["sell_to"],
            "pair": f"{r['buy_from']} → {r['sell_to']}",
            "count": int(r["count"]),
            "total_net_profit_usd": round(float(r["total_net_profit_usd"] or 0), 2),
            "total_fees_usd": round(float(r["total_fees_usd"] or 0), 2),
            "avg_apr_pct": round(float(r["avg_apr_pct"] or 0), 1),
            "best_net_profit_usd": round(float(r["best_net_profit_usd"] or 0), 2),
        }
        for r in rows
    ]


_SORT_COLS = {
    "detected_at": "detected_at",
    "apr_pct": "apr_pct",
    "net_return_pct": "net_return_pct",
    "net_profit_usd": "net_profit_usd",
    "buy_premium_usd": "buy_premium_usd",
    "fees_usd": "fees_usd",
}


@router.get("", response_model=list[OpportunityResponse])
async def list_opportunities(
    status: OpportunityStatus | None = None,
    min_apr: float | None = None,
    min_profit: float | None = None,
    symbol: str | None = None,
    buy_from: str | None = None,
    sell_to: str | None = None,
    days: int | None = None,
    network: Network | None = "mainnet",
    sort_by: OpportunitySortBy = "detected_at",
    sort_dir: SortDirection = "desc",
    limit: int = Query(default=100, le=1000),
    offset: int = Query(default=0, ge=0),
) -> list[Opportunity]:
    from datetime import UTC, datetime, timedelta

    import sqlalchemy as sa

    col_expr = sa.text(f"{_SORT_COLS[sort_by]} {sort_dir.upper()}")
    stmt = select(Opportunity).order_by(col_expr).limit(limit).offset(offset)
    if status is not None:
        stmt = stmt.where(Opportunity.status == status)
    if min_apr is not None:
        stmt = stmt.where(Opportunity.apr_pct >= min_apr)
    if min_profit is not None:
        stmt = stmt.where(Opportunity.net_profit_usd >= min_profit)
    if symbol is not None:
        stmt = stmt.where(Opportunity.symbol == symbol)
    if buy_from is not None:
        stmt = stmt.where(Opportunity.buy_from == buy_from)
    if sell_to is not None:
        stmt = stmt.where(Opportunity.sell_to == sell_to)
    if days is not None:
        since = datetime.now(UTC) - timedelta(days=days)
        stmt = stmt.where(sa.cast(Opportunity.detected_at, sa.DateTime(timezone=True)) >= since)
    if network is not None:
        stmt = stmt.where(Opportunity.network == network)
    async with get_session() as sess:
        return list((await sess.execute(stmt)).scalars())


@router.get(
    "/{opp_id}",
    response_model=OpportunityResponse,
    responses={404: {"model": ErrorResponse}},
)
async def get_opportunity(opp_id: int) -> Opportunity:
    async with get_session() as sess:
        row = (
            await sess.execute(select(Opportunity).where(Opportunity.id == opp_id))
        ).scalar_one_or_none()
    if row is None:
        raise HTTPException(404, "not found")
    return row