# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "additionalProperties": false,
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          }
        },
        "required": [
          "detail"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "IsoDatetime": {
        "format": "date-time",
        "type": "string"
      },
      "Mode": {
        "enum": [
          "live",
          "paper",
          "backtest"
        ],
        "title": "Mode",
        "type": "string"
      },
      "Network": {
        "enum": [
          "mainnet",
          "testnet"
        ],
        "type": "string"
      },
      "OpportunityResponse": {
        "additionalProperties": false,
        "properties": {
          "apr_pct": {
            "title": "Apr Pct",
            "type": "number"
          },
          "buy_from": {
            "title": "Buy From",
            "type": "string"
          },
          "buy_premium_usd": {
            "title": "Buy Premium Usd",
            "type": "number"
          },
          "capital_required_usd": {
            "title": "Capital Required Usd",
            "type": "number"
          },
          "days_to_expiry": {
            "readOnly": true,
            "title": "Days To Expiry",
            "type": "number"
          },
          "detected_at": {
            "$ref": "#/components/schemas/IsoDatetime"
          },
          "estimated_short_margin_usd": {
            "title": "Estimated Short Margin Usd",
            "type": "number"
          },
          "expiry": {
            "$ref": "#/components/schemas/IsoDatetime"
          },
          "fees_usd": {
            "title": "Fees Usd",
            "type": "number"
          },
          "gross_profit_usd": {
            "title": "Gross Profit Usd",
            "type": "number"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "instrument": {
            "title": "Instrument",
            "type": "string"
          },
          "mode": {
            "$ref": "#/components/schemas/Mode"
          },
          "net_profit_usd": {
            "title": "Net Profit Usd",
            "type": "number"
          },
          "net_return_pct": {
            "title": "Net Return Pct",
            "type": "number"
          },
          "network": {
            "$ref": "#/components/schemas/Network"
          },
          "option_type": {
            "enum": [
              "C",
              "P"
            ],
            "title": "Option Type",
            "type": "string"
          },
          "price_spread_pct": {
            "title": "Price Spread Pct",
            "type": "number"
          },
          "rejection_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rejection Reason"
          },
          "sell_premium_usd": {
            "title": "Sell Premium Usd",
            "type": "number"
          },
          "sell_to": {
            "title": "Sell To",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/OpportunityStatus"
          },
          "strike": {
            "title": "Strike",
            "type": "number"
          },
          "symbol": {
            "title": "Symbol",
            "type": "string"
          },
          "top_ask": {
            "title": "Top Ask",
            "type": "number"
          },
          "top_bid": {
            "title": "Top Bid",
            "type": "number"
          },
          "tradeable_size": {
            "title": "Tradeable Size",
            "type": "number"
          },
          "verified_apr_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Verified Apr Pct"
          },
          "verified_buy_limit": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Verified Buy Limit"
          },
          "verified_buy_premium_usd": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Verified Buy Premium Usd"
          },
          "verified_capital_required_usd": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Verified Capital Required Usd"
          },
          "verified_estimated_short_margin_usd": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Verified Estimated Short Margin Usd"
          },
          "verified_fees_usd": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Verified Fees Usd"
          },
          "verified_gross_profit_usd": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Verified Gross Profit Usd"
          },
          "verified_net_profit_usd": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Verified Net Profit Usd"
          },
          "verified_net_return_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Verified Net Return Pct"
          },
          "verified_sell_limit": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Verified Sell Limit"
          },
          "verified_sell_premium_usd": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Verified Sell Premium Usd"
          },
          "verified_tradeable_size": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Verified Tradeable Size"
          },
          "walked_ask": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Walked Ask"
          },
          "walked_bid": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Walked Bid"
          }
        },
        "required": [
          "id",
          "detected_at",
          "mode",
          "network",
          "instrument",
          "symbol",
          "expiry",
          "strike",
          "option_type",
          "buy_from",
          "sell_to",
          "top_ask",
          "top_bid",
          "walked_ask",
          "walked_bid",
          "tradeable_size",
          "buy_premium_usd",
          "sell_premium_usd",
          "estimated_short_margin_usd",
          "capital_required_usd",
          "gross_profit_usd",
          "fees_usd",
          "net_profit_usd",
          "price_spread_pct",
          "net_return_pct",
          "apr_pct",
          "verified_buy_limit",
          "verified_sell_limit",
          "verified_tradeable_size",
          "verified_buy_premium_usd",
          "verified_sell_premium_usd",
          "verified_estimated_short_margin_usd",
          "verified_capital_required_usd",
          "verified_gross_profit_usd",
          "verified_fees_usd",
          "verified_net_profit_usd",
          "verified_net_return_pct",
          "verified_apr_pct",
          "status",
          "rejection_reason",
          "days_to_expiry"
        ],
        "title": "OpportunityResponse",
        "type": "object"
      },
      "OpportunityStatsResponse": {
        "additionalProperties": false,
        "properties": {
          "avg_apr_pct": {
            "title": "Avg Apr Pct",
            "type": "number"
          },
          "best_net_profit_usd": {
            "title": "Best Net Profit Usd",
            "type": "number"
          },
          "buy_from": {
            "title": "Buy From",
            "type": "string"
          },
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "pair": {
            "title": "Pair",
            "type": "string"
          },
          "sell_to": {
            "title": "Sell To",
            "type": "string"
          },
          "total_fees_usd": {
            "title": "Total Fees Usd",
            "type": "number"
          },
          "total_net_profit_usd": {
            "title": "Total Net Profit Usd",
            "type": "number"
          }
        },
        "required": [
          "buy_from",
          "sell_to",
          "pair",
          "count",
          "total_net_profit_usd",
          "total_fees_usd",
          "avg_apr_pct",
          "best_net_profit_usd"
        ],
        "title": "OpportunityStatsResponse",
        "type": "object"
      },
      "OpportunityStatus": {
        "enum": [
          "PENDING",
          "APPROVED",
          "REJECTED",
          "EXECUTED",
          "EXPIRED"
        ],
        "title": "OpportunityStatus",
        "type": "string"
      },
...
      "TradeResponse": {
        "additionalProperties": false,
        "properties": {
          "buy_exchange": {
            "title": "Buy Exchange",
            "type": "string"
          },
          "buy_fill_price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Buy Fill Price"
          },
          "buy_fill_size": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Buy Fill Size"
          },
          "closed_at": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/IsoDatetime"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "fees_usd": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fees Usd"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "mode": {
            "$ref": "#/components/schemas/Mode"
          },
          "net_pnl_usd": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Net Pnl Usd"
          },
          "opened_at": {
            "$ref": "#/components/schemas/IsoDatetime"
          },
          "opportunity_id": {
            "title": "Opportunity Id",
            "type": "integer"
          },
          "requested_size": {
            "title": "Requested Size",
            "type": "number"
          },
          "sell_exchange": {
            "title": "Sell Exchange",
            "type": "string"
          },
          "sell_fill_price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sell Fill Price"
          },
          "sell_fill_size": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sell Fill Size"
          },
          "slippage_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Slippage Pct"
          },
          "status": {
            "$ref": "#/components/schemas/TradeStatus"
          }
        },
        "required": [
          "id",
          "opportunity_id",
          "opened_at",
          "closed_at",
          "mode",
          "status",
          "buy_exchange",
          "sell_exchange",
          "requested_size",
          "buy_fill_price",
          "buy_fill_size",
          "sell_fill_price",
          "sell_fill_size",
          "net_pnl_usd",
          "slippage_pct",
          "fees_usd",
          "error"
        ],
        "title": "TradeResponse",
        "type": "object"
      },
      "TradeStatus": {
        "enum": [
          "PLACING",
          "LEG1_FILLED",
          "LEG2_FILLED",
          "FILLED",
          "HEDGING",
          "HEDGED",
          "STUCK",
          "FAILED"
        ],
        "title": "TradeStatus",
        "type": "string"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "option-arb API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/opportunities": {
      "get": {
        "operationId": "list_opportunities_api_opportunities_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/OpportunityStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "min_apr",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Min Apr"
            }
          },
          {
            "in": "query",
            "name": "min_profit",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Min Profit"
            }
          },
          {
            "in": "query",
            "name": "symbol",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Symbol"
            }
          },
          {
            "in": "query",
            "name": "buy_from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Buy From"
            }
          },
          {
            "in": "query",
            "name": "sell_to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sell To"
            }
          },
          {
            "in": "query",
            "name": "days",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Days"
            }
          },
          {
            "in": "query",
            "name": "network",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Network"
                },
                {
                  "type": "null"
                }
              ],
              "default": "mainnet",
              "title": "Network"
            }
          },
          {
            "in": "query",
            "name": "sort_by",
            "required": false,
            "schema": {
              "default": "detected_at",
              "enum": [
                "detected_at",
                "apr_pct",
                "net_return_pct",
                "net_profit_usd",
                "buy_premium_usd",
                "fees_usd"
              ],
              "title": "Sort By",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "sort_dir",
            "required": false,
            "schema": {
              "default": "desc",
              "enum": [
                "asc",
                "desc"
              ],
              "title": "Sort Dir",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 1000,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/OpportunityResponse"
                  },
                  "title": "Response List Opportunities Api Opportunities Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Opportunities",
        "tags": [
          "opportunities"
        ]
      }
    },
    "/api/opportunities/stats": {
      "get": {
        "description": "Opportunit\u00e9s agr\u00e9g\u00e9es par paire d'exchanges sur N jours.",
        "operationId": "opportunity_stats_api_opportunities_stats_get",
        "parameters": [
          {
            "in": "query",
            "name": "days",
            "required": false,
            "schema": {
              "default": 30,
              "maximum": 365,
              "minimum": 1,
              "title": "Days",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "symbol",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Symbol"
            }
          },
          {
            "in": "query",
            "name": "network",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Network"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Network"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/OpportunityStatsResponse"
                  },
                  "title": "Response Opportunity Stats Api Opportunities Stats Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Opportunity Stats",
        "tags": [
          "opportunities"
        ]
      }
    },
    "/api/opportunities/{opp_id}": {
      "get": {
        "operationId": "get_opportunity_api_opportunities__opp_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "opp_id",
            "required": true,
            "schema": {
              "title": "Opp Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpportunityResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Opportunity",
        "tags": [
          "opportunities"
        ]
      }
    },
...
    "/api/trades": {
      "get": {
        "operationId": "list_trades_api_trades_get",
        "parameters": [
          {
            "in": "query",
            "name": "mode",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Mode"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Mode"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/TradeStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 500,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TradeResponse"
                  },
                  "title": "Response List Trades Api Trades Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Trades",
        "tags": [
          "trades"
        ]
      }
    },
    "/api/trades/{trade_id}": {
      "get": {
        "operationId": "get_trade_api_trades__trade_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "trade_id",
            "required": true,
            "schema": {
              "title": "Trade Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TradeDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Trade",
        "tags": [
          "trades"
        ]
      }
    },
...
  }
}